+++ NEWS.md
@@ -2,9 +2,11 @@
 peacesciencer 0.2.0
 ---------------------------------------------------------------------
 
-- Fix stupid mistake in README.
+- Fix stupid mistake(s) in README.
 - File compression for the data files.
 - Add Correlates of War intergovernmental organizations data.
+- Add function for adding Correlates of War intergovernmental organizations data.
+- Add minimum distance data
 - Add citations data frame for assistance in properly citing things.
 - Add `create_statedays()` function.